mod 10=9 forcing logging 16 floats : 100Hz at 999ms, 10Hz at 622ms

+++ AeroQuad/AeroQuad.ino
@@ -1585,7 +1585,7 @@ void loop () {
 //          }
 //          LogValueSpace((float)batteryData[0].voltage/100.0);
           for (byte motor = 0; motor < 16; motor++) {
-            LogValueSpace((short)1);
+            LogValueSpace((float)1);
           }
         
           LOG_SERIAL.println();
